Move version & Ko-fi from sidebar footer to top bar, remove location summary
- Relocate version display and Ko-fi tip button from SiderFooter to Header
- Show compact (icon-only) version on mobile, full text on desktop
- Remove "X Spools in Y locations" summary line from locations page
- Fix React hooks order violation in LocationContainer (useState after early return)
